refactor(bigtable): move some definitions around
